editoast: move train schedule schema files to paced train
first step before removing schema/train_schedule.rs
