+++ editoast/core_client/src/simulation.rs
@@ -10,14 +10,14 @@ use common::units::quantities::Mass;
 use common::units::quantities::Time;
 use common::units::quantities::Velocity;
 use educe::Educe;
+use schemas::paced_train::Comfort;
+use schemas::paced_train::Distribution;
+use schemas::paced_train::MarginValue;
+use schemas::paced_train::ReceptionSignal;
+use schemas::paced_train::TrainScheduleOptions;
 use schemas::rolling_stock::EffortCurves;
 use schemas::rolling_stock::EtcsBrakeParams;
 use schemas::rolling_stock::RollingResistance;
-use schemas::train_schedule::Comfort;
-use schemas::train_schedule::Distribution;
-use schemas::train_schedule::MarginValue;
-use schemas::train_schedule::ReceptionSignal;
-use schemas::train_schedule::TrainScheduleOptions;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
