+++ editoast/core_client/src/stdcm.rs
@@ -3,9 +3,9 @@ use std::collections::HashSet;
 use chrono::DateTime;
 use chrono::Utc;
 use schemas::infra::TrackOffset;
+use schemas::paced_train::Comfort;
+use schemas::paced_train::MarginValue;
 use schemas::rolling_stock::LoadingGaugeType;
-use schemas::train_schedule::Comfort;
-use schemas::train_schedule::MarginValue;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
