+++ editoast/schemas/src/paced_train/allowance.rs
