+++ editoast/schemas/src/paced_train/comfort.rs
