+++ editoast/schemas/src/paced_train/path_item.rs
