+++ editoast/schemas/src/paced_train/power_restriction_item.rs
