+++ editoast/schemas/src/paced_train/rjs_power_restriction_range.rs
