+++ editoast/schemas/src/paced_train/schedule_item.rs
