+++ editoast/schemas/src/paced_train/train_schedule_options.rs
