+++ editoast/src/models/fixtures.rs
@@ -24,34 +24,34 @@ use schemas::fixtures::simple_modified_exception_with_change_groups;
 use schemas::infra::InfraObject;
 use schemas::infra::RailJson;
 use schemas::infra::TrackOffset;
+use schemas::paced_train::Comfort;
 use schemas::paced_train::ConstraintDistributionChangeGroup;
+use schemas::paced_train::Distribution;
 use schemas::paced_train::ExceptionType;
 use schemas::paced_train::InitialSpeedChangeGroup;
 use schemas::paced_train::LabelsChangeGroup;
+use schemas::paced_train::MarginValue;
+use schemas::paced_train::Margins;
+use schemas::paced_train::OperationalPointPartReference;
+use schemas::paced_train::OperationalPointReference;
 use schemas::paced_train::OptionsChangeGroup;
 use schemas::paced_train::Paced;
 use schemas::paced_train::PacedTrain;
 use schemas::paced_train::PacedTrainException;
 use schemas::paced_train::PathAndScheduleChangeGroup;
+use schemas::paced_train::PathItem;
+use schemas::paced_train::PathItemLocation;
 use schemas::paced_train::RollingStockCategoryChangeGroup;
 use schemas::paced_train::RollingStockChangeGroup;
+use schemas::paced_train::ScheduleItem;
 use schemas::paced_train::SpeedLimitTagChangeGroup;
 use schemas::paced_train::StartTimeChangeGroup;
 use schemas::paced_train::TrainNameChangeGroup;
+use schemas::paced_train::TrainScheduleOptions;
 use schemas::primitives::Identifier;
 use schemas::primitives::NonBlankString;
 use schemas::primitives::OSRDObject;
 use schemas::rolling_stock::SubCategoryColor;
-use schemas::train_schedule::Comfort;
-use schemas::train_schedule::Distribution;
-use schemas::train_schedule::MarginValue;
-use schemas::train_schedule::Margins;
-use schemas::train_schedule::OperationalPointPartReference;
-use schemas::train_schedule::OperationalPointReference;
-use schemas::train_schedule::PathItem;
-use schemas::train_schedule::PathItemLocation;
-use schemas::train_schedule::ScheduleItem;
-use schemas::train_schedule::TrainScheduleOptions;
 
 use crate::infra_cache::operation::create::apply_create_operation;
 use crate::models;
