+++ editoast/src/views/search.rs
@@ -218,11 +218,11 @@ use diesel_async::RunQueryDsl;
 use editoast_derive::EditoastError;
 use editoast_derive::Search;
 use editoast_derive::SearchConfigStore;
-use schemas::train_schedule::Margins;
-use schemas::train_schedule::PathItem;
-use schemas::train_schedule::PowerRestrictionItem;
-use schemas::train_schedule::ScheduleItem;
-use schemas::train_schedule::TrainScheduleOptions;
+use schemas::paced_train::Margins;
+use schemas::paced_train::PathItem;
+use schemas::paced_train::PowerRestrictionItem;
+use schemas::paced_train::ScheduleItem;
+use schemas::paced_train::TrainScheduleOptions;
 use search::SearchConfigStore as _;
 use search::SearchError;
 use search::query_into_sql;
