+++ editoast/src/views/timetable/paced_train.rs
@@ -20,10 +20,10 @@ use editoast_models::round_trips::PacedTrainRoundTrips;
 use itertools::Itertools;
 use itertools::izip;
 use reqwest::StatusCode;
+use schemas::paced_train::OperationalPointPartReference;
 use schemas::paced_train::PacedTrain;
+use schemas::paced_train::PathItemLocation;
 use schemas::primitives::TimeWindow;
-use schemas::train_schedule::OperationalPointPartReference;
-use schemas::train_schedule::PathItemLocation;
 use serde::Deserialize;
 use serde::Serialize;
 use thiserror::Error;
@@ -1473,17 +1473,17 @@ mod tests {
     use schemas::fixtures::simple_created_exception_with_change_groups;
     use schemas::fixtures::simple_modified_exception_with_change_groups;
     use schemas::infra::Direction;
+    use schemas::paced_train::Comfort;
     use schemas::paced_train::ExceptionType;
     use schemas::paced_train::InitialSpeedChangeGroup;
     use schemas::paced_train::Paced;
     use schemas::paced_train::PacedTrain;
     use schemas::paced_train::PacedTrainException;
+    use schemas::paced_train::PathItem;
     use schemas::paced_train::RollingStockChangeGroup;
+    use schemas::paced_train::ScheduleItem;
     use schemas::paced_train::TrainNameChangeGroup;
     use schemas::rolling_stock::TrainCategory;
-    use schemas::train_schedule::Comfort;
-    use schemas::train_schedule::PathItem;
-    use schemas::train_schedule::ScheduleItem;
     use schemas::train_schedule::TrainSchedule;
     use serde_json::json;
