+++ editoast/src/views/timetable/simulation.rs
@@ -14,10 +14,10 @@ use core_client::simulation::SimulationScheduleItem;
 use core_client::simulation::SpeedLimitProperties;
 use database::DbConnection;
 use itertools::Itertools;
-use schemas::train_schedule::Margins;
-use schemas::train_schedule::PathItem;
-use schemas::train_schedule::PowerRestrictionItem;
-use schemas::train_schedule::ScheduleItem;
+use schemas::paced_train::Margins;
+use schemas::paced_train::PathItem;
+use schemas::paced_train::PowerRestrictionItem;
+use schemas::paced_train::ScheduleItem;
 use schemas::train_schedule::TrainScheduleLike;
 use serde::Deserialize;
 use serde::Serialize;
