+++ editoast/src/views/timetable/stdcm.rs
@@ -19,11 +19,12 @@ use database::DbConnectionPoolV2;
 use editoast_derive::EditoastError;
 use request::Request;
 use request::convert_steps;
+use schemas::paced_train::MarginValue;
+use schemas::paced_train::Margins;
+use schemas::paced_train::ReceptionSignal;
+use schemas::paced_train::ScheduleItem;
 use schemas::primitives::PositiveDuration;
-use schemas::train_schedule::MarginValue;
-use schemas::train_schedule::Margins;
-use schemas::train_schedule::ReceptionSignal;
-use schemas::train_schedule::ScheduleItem;
+
 use schemas::train_schedule::TrainSchedule;
 use serde::Deserialize;
 use serde::Serialize;
@@ -483,11 +484,11 @@ mod tests {
     use rstest::rstest;
     use schemas::fixtures::simple_rolling_stock;
     use schemas::fixtures::towed_rolling_stock;
+    use schemas::paced_train::Comfort;
+    use schemas::paced_train::OperationalPointPartReference;
+    use schemas::paced_train::OperationalPointReference;
+    use schemas::paced_train::PathItemLocation;
     use schemas::rolling_stock::RollingResistance;
-    use schemas::train_schedule::Comfort;
-    use schemas::train_schedule::OperationalPointPartReference;
-    use schemas::train_schedule::OperationalPointReference;
-    use schemas::train_schedule::PathItemLocation;
     use serde_json::json;
     use std::str::FromStr;
     use uom::si::length::Length;
